perf: Use string interpolation instead of set
After playing around in jsperf and reading around on available information regarding performance, I decided to switch over to string operations instead of list operations to increase the overall performance of the utility function.

The result is sadly that the code is not as sexy as it is with the use of Set, and the introduction of string concatenation/interpolation makes duplication prevention into a tedious task which will most likely result in worse then original performance.

I, therefore, decided to discard duplication prevention in favor of performance, as the operation will be ran at a high frequency when used to create component classnames. It is now up to the programmer to prevent duplicate classes from apearing.

TLDR; Performance boost by changing from using set to direct string operations and duplicate classes are now a possibility.

eks:

Before: cookedNames("foo", "bar", "foo", "bar") => "foo bar";
Now: cookedNames("foo", "bar", "foo", "bar") => "foo bar foo bar";