perf: Use native operations for type checking
Use native methods such as typeof and Array.isArray to retrieve the argument type instead of the string/regex combination used earlier to increase the overall performance of the utility function.

Additionally, switched to direct recursive function calls instead if apply, as the direct method seems to be alot more performant on newer browsers compared to both .apply and .call.